Move the instance type and dimensions management to a new ami-product subcommand
