+++ aws_mp_utils/ami_product_instance_types.py
@@ -21,20 +21,19 @@
 
 import boto3
 import jmespath
-import json
 
 
-def get_available_dimensions(
+def get_available_instance_types(
     client: boto3.client,
-    offer_id: str,
+    product_id: str,
     catalog: str = 'AWSMarketplace'
 ) -> list[str]:
     """
-    Lists the available dimensions for the given offer.
+    Lists the available instance types for the given product.
     """
     entity = client.describe_entity(
         Catalog=catalog,
-        EntityId=offer_id
+        EntityId=product_id
     )
 
     """
@@ -72,50 +71,61 @@ def get_available_dimensions(
     """
 
     details = entity['DetailsDocument']
-    query = "Dimensions"
-    dimensions = jmespath.search(query, details)
+    query = "Versions[].Sources[].Compatibility.AvailableInstanceTypes[]"
+    instance_types = jmespath.search(query, details)
 
-    if dimensions is None:
+    if instance_types is None:
         return []
-    return sorted(dimensions, key=lambda x: x['Name'])
+    return sorted(list(set(instance_types)))
 
 
-def create_restrict_dimensions_change_doc(
-    offer_id: str,
-    details_document: str,
+def create_restrict_instance_types_change_doc(
+    product_id: str,
+    instance_types: [str],
 ) -> dict:
     """
-    Creates an update offer request dictionary to restrict dimensions.
+    Creates an update product request dictionary to restrict instance types.
 
-    :param offer_id: The unique identifier of the offer in the AWS Marketplace.
-    :param details_document: A JSON formatted string containing the details
-        document for restricting the offer dimensions.
+    :param product_id: The unique identifier of the product in the AWS
+    Marketplace.
+    :param instance_types: A list of instance types for restriction in the
+    ami product.
     """
     data = {
-        'ChangeType': "RestrictDimensions",
+        'ChangeType': "RestrictInstanceTypes",
         'Entity': {
-            'Type': 'Offer@1.0',
-            'Identifier': offer_id
+            'Type': 'AmiProduct@1.0',
+            'Identifier': product_id
         },
-        'DetailsDocument': json.loads(details_document)
+        'DetailsDocument': {
+            'InstanceTypes': instance_types
+        }
     }
     return data
 
 
-def create_add_dimensions_change_doc(
-    offer_id: str,
-    details_document: str,
+def create_add_instance_types_change_doc(
+    product_id: str,
+    instance_types: [str],
 ) -> dict:
     """
-    Creates an update offer request dictionary to add dimensions.
+    Creates an addInstanceTypes update product request dictionary to add
+    available instance types.
+
+    :param product_id: The unique identifier of the product in the AWS
+    Marketplace.
+    :param instance_types: A list of instance types for addition in the
+        product.
     """
 
     data = {
-        'ChangeType': "AddDimensions",
+        'ChangeType': "AddInstanceTypes",
         'Entity': {
-            'Type': 'Offer@1.0',
-            'Identifier': offer_id
+            'Type': 'AmiProduct@1.0',
+            'Identifier': product_id
         },
-        'DetailsDocument': json.loads(details_document)
+        'DetailsDocument': {
+            'InstanceTypes': instance_types
+        }
     }
     return data
